fetch/metadata/generated/script-module-import-static-sec-fetch-site.https.sub.html

<!DOCTYPE html>
<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta charset="utf-8">
  <title>Fetch Metadata - Sec-Fetch-Site - redirection from HTTPS</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(moduleId) {
    const script = document.createElement('script');
    script.setAttribute('type', 'module');
    script.setAttribute(
      'src', '/fetch/metadata/resources/es-module.sub.js?moduleId=' + moduleId
    );

    return new Promise((resolve, reject) => {
        script.onload = resolve;
        script.onerror = () => reject('Failed to load script');
        document.body.appendChild(script);
      })
      .then(() => script.remove());
  }

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py?key={{uuid()}}';
    const moduleId = 'https://{{host}}:{{ports[https][0]}}/fetch/api/resources/redirect.py?location=https://{{hosts[alt][www]}}:{{ports[https][0]}}/fetch/api/resources/redirect.py?location=https://{{host}}:{{ports[https][0]}}' +
      encodeURIComponent(path + '&mime=application/javascript');

    return induceRequest(moduleId)
      .then(() => fetch(path + '&retrieve'))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - Same-Origin -> Cross-Site -> Same-Origin redirect');
  </script>
</html>
